General improvements
- CustomUserDetailsService remove
- project structure rework
- AuthService rip into 2 services
- deprecated settings remove
