+++ backend/src/main/java/net/archasmiel/skufapi/Application.java
@@ -4,9 +4,10 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class SkufApiApplication {
+public class Application {
 
   public static void main(String[] args) {
-    SpringApplication.run(SkufApiApplication.class, args);
+    SpringApplication.run(Application.class, args);
   }
+
 }
