+++ backend/src/main/java/net/archasmiel/skufapi/api/request/ApiRequest.java
@@ -1,0 +1,4 @@
+package net.archasmiel.skufapi.api.request;
+
+public interface ApiRequest {
+}
