+++ backend/src/main/java/net/archasmiel/skufapi/api/request/auth/GoogleAuthRequest.java
@@ -1,19 +1,19 @@
-package net.archasmiel.skufapi.dto.auth;
+package net.archasmiel.skufapi.api.request.auth;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import net.archasmiel.skufapi.dto.IRequest;
+import net.archasmiel.skufapi.api.request.ApiRequest;
 
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 @Schema(description = "Sign in request via Google")
-public class GoogleAuthRequest implements IRequest {
+public class GoogleAuthRequest implements ApiRequest {
 
-  @Schema(description = "GoogleId token", example = "eyJhbGciOiJIUzUxMiJ9...")
-  private String token;
+    @Schema(description = "GoogleId token", example = "eyJhbGciOiJIUzUxMiJ9...")
+    private String token;
 }
