+++ backend/src/main/java/net/archasmiel/skufapi/api/response/ApiResponse.java
@@ -1,0 +1,4 @@
+package net.archasmiel.skufapi.api.response;
+
+public interface ApiResponse {
+}
