+++ backend/src/main/java/net/archasmiel/skufapi/api/response/auth/JwtAuthResponse.java
@@ -1,19 +1,19 @@
-package net.archasmiel.skufapi.dto.auth;
+package net.archasmiel.skufapi.api.response.auth;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import net.archasmiel.skufapi.dto.IResponse;
+import net.archasmiel.skufapi.api.response.ApiResponse;
 
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 @Schema(description = "Response with Jwt token")
-public class JwtAuthResponse implements IResponse {
+public class JwtAuthResponse implements ApiResponse {
 
-  @Schema(description = "Jwt token", example = "eyJhbGciOiJIUzUxMiJ9...")
-  private String jwtToken;
+    @Schema(description = "Jwt token", example = "eyJhbGciOiJIUzUxMiJ9...")
+    private String jwtToken;
 }
