+++ backend/src/main/java/net/archasmiel/skufapi/api/response/auth/MeResponse.java
@@ -1,0 +1,22 @@
+package net.archasmiel.skufapi.api.response.auth;
+
+import io.swagger.v3.oas.annotations.media.Schema;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import net.archasmiel.skufapi.api.response.ApiResponse;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@Schema(description = "Response with Jwt token")
+public class MeResponse implements ApiResponse {
+
+    @Schema(description = "Jwt token", example = "eyJhbGciOiJIUzUxMiJ9...")
+    private String username;
+
+    @Schema(description = "Jwt token", example = "eyJhbGciOiJIUzUxMiJ9...")
+    private String roles;
+}
