+++ backend/src/main/java/net/archasmiel/skufapi/config/EnvConfig.java
@@ -5,4 +5,5 @@ import org.springframework.context.annotation.PropertySource;
 
 @Configuration
 @PropertySource("classpath:.env")
-public class EnvConfig {}
+public class EnvConfig {
+}
