+++ backend/src/main/java/net/archasmiel/skufapi/config/GoogleClientConfig.java
@@ -3,26 +3,27 @@ package net.archasmiel.skufapi.config;
 import com.google.api.client.googleapis.auth.oauth2.GoogleIdTokenVerifier;
 import com.google.api.client.http.javanet.NetHttpTransport;
 import com.google.api.client.json.gson.GsonFactory;
-import java.util.Collections;
 import lombok.Getter;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.stereotype.Component;
 
+import java.util.Collections;
+
 @Getter
 @Component
 public class GoogleClientConfig {
 
-  @Value("${spring.security.oauth2.client.registration.google.client-id}")
-  private String googleClientId;
+    @Value("${spring.security.oauth2.client.registration.google.client-id}")
+    private String googleClientId;
 
-  @Value("${spring.security.oauth2.client.registration.google.client-id}")
-  private String googleClientSecret;
+    @Value("${spring.security.oauth2.client.registration.google.client-id}")
+    private String googleClientSecret;
 
-  @Bean
-  public GoogleIdTokenVerifier googleIdTokenVerifier() {
-    return new GoogleIdTokenVerifier.Builder(new NetHttpTransport(), new GsonFactory())
-        .setAudience(Collections.singletonList(googleClientId))
-        .build();
-  }
+    @Bean
+    public GoogleIdTokenVerifier googleIdTokenVerifier() {
+        return new GoogleIdTokenVerifier.Builder(new NetHttpTransport(), new GsonFactory())
+                .setAudience(Collections.singletonList(googleClientId))
+                .build();
+    }
 }
