+++ backend/src/main/java/net/archasmiel/skufapi/config/JwtAuthFilter.java
@@ -4,8 +4,6 @@ import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.util.List;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import net.archasmiel.skufapi.service.JwtService;
@@ -19,69 +17,69 @@ import org.springframework.security.web.authentication.WebAuthenticationDetailsS
 import org.springframework.stereotype.Component;
 import org.springframework.web.filter.OncePerRequestFilter;
 
+import java.io.IOException;
+import java.util.List;
+
 @Component
 @RequiredArgsConstructor
 public class JwtAuthFilter extends OncePerRequestFilter {
 
-  @Value("${key.secret.frontend}")
-  private String keySecret;
+    public static final String BEARER_PREFIX = "Bearer ";
+    public static final String HEADER_NAME = "Authorization";
 
-  public static final String BEARER_PREFIX = "Bearer ";
-  public static final String HEADER_NAME = "Authorization";
+    private final JwtService jwtService;
+    private final UserService userService;
 
-  private final JwtService jwtService;
-  private final UserService userService;
+    // TODO
+    private static final List<String> AUTH_WHITELIST =
+            List.of("/api/auth/login", "/api/auth/register", "/api/auth/signup");
 
-  // TODO
-  private static final List<String> AUTH_WHITELIST =
-      List.of("/api/auth/login", "/api/auth/register", "/api/auth/signup");
-
-  @Override
-  protected void doFilterInternal(
-      @NonNull HttpServletRequest request,
-      @NonNull HttpServletResponse response,
-      @NonNull FilterChain filterChain)
-      throws ServletException, IOException {
-    if (AUTH_WHITELIST.contains(request.getServletPath())) {
-      filterChain.doFilter(request, response);
-      return;
-    }
-
-    handleStandardJwt(request, response, filterChain);
-  }
-
-  private void handleStandardJwt(
-      HttpServletRequest request, HttpServletResponse response, FilterChain filterChain)
-      throws IOException, ServletException {
-    var authHeader = request.getHeader(HEADER_NAME);
-    if (StringUtils.isEmpty(authHeader) || !StringUtils.startsWith(authHeader, BEARER_PREFIX)) {
-      filterChain.doFilter(request, response);
-      return;
-    }
-
-    var jwt = authHeader.substring(BEARER_PREFIX.length());
-    try {
-      var username = jwtService.extractUsername(jwt);
-
-      if (StringUtils.isNotEmpty(username)
-          && SecurityContextHolder.getContext().getAuthentication() == null) {
-        UserDetails userDetails = userService.userDetailsService().loadUserByUsername(username);
-
-        if (jwtService.isTokenValid(jwt, userDetails)) {
-          setAuthentication(userDetails, request);
+    @Override
+    protected void doFilterInternal(
+            @NonNull HttpServletRequest request,
+            @NonNull HttpServletResponse response,
+            @NonNull FilterChain filterChain)
+            throws ServletException, IOException {
+        if (AUTH_WHITELIST.contains(request.getServletPath())) {
+            filterChain.doFilter(request, response);
+            return;
         }
-        filterChain.doFilter(request, response);
-      }
 
-    } catch (Exception e) {
-      response.sendError(HttpServletResponse.SC_UNAUTHORIZED, "Invalid JWT");
+        handleStandardJwt(request, response, filterChain);
     }
-  }
 
-  private void setAuthentication(UserDetails userDetails, HttpServletRequest request) {
-    UsernamePasswordAuthenticationToken authToken =
-        new UsernamePasswordAuthenticationToken(userDetails, null, userDetails.getAuthorities());
-    authToken.setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
-    SecurityContextHolder.getContext().setAuthentication(authToken);
-  }
+    private void handleStandardJwt(
+            HttpServletRequest request, HttpServletResponse response, FilterChain filterChain)
+            throws IOException, ServletException {
+        var authHeader = request.getHeader(HEADER_NAME);
+        if (StringUtils.isEmpty(authHeader) || !StringUtils.startsWith(authHeader, BEARER_PREFIX)) {
+            filterChain.doFilter(request, response);
+            return;
+        }
+
+        var jwt = authHeader.substring(BEARER_PREFIX.length());
+        try {
+            var username = jwtService.extractUsername(jwt);
+
+            if (StringUtils.isNotEmpty(username)
+                    && SecurityContextHolder.getContext().getAuthentication() == null) {
+                UserDetails userDetails = userService.userDetailsService().loadUserByUsername(username);
+
+                if (jwtService.isTokenValid(jwt, userDetails)) {
+                    setAuthentication(userDetails, request);
+                }
+                filterChain.doFilter(request, response);
+            }
+
+        } catch (Exception e) {
+            response.sendError(HttpServletResponse.SC_UNAUTHORIZED, "Invalid JWT");
+        }
+    }
+
+    private void setAuthentication(UserDetails userDetails, HttpServletRequest request) {
+        UsernamePasswordAuthenticationToken authToken =
+                new UsernamePasswordAuthenticationToken(userDetails, null, userDetails.getAuthorities());
+        authToken.setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
+        SecurityContextHolder.getContext().setAuthentication(authToken);
+    }
 }
