+++ backend/src/main/java/net/archasmiel/skufapi/domain/enums/Role.java
@@ -1,0 +1,6 @@
+package net.archasmiel.skufapi.domain.enums;
+
+public enum Role {
+    ROLE_USER,
+    ROLE_ADMIN,
+}
