--- backend/src/main/java/net/archasmiel/skufapi/dto/IRequest.java
@@ -1,3 +1,0 @@
-package net.archasmiel.skufapi.dto;
-
-public interface IRequest {}
