--- backend/src/main/java/net/archasmiel/skufapi/dto/IResponse.java
@@ -1,3 +1,0 @@
-package net.archasmiel.skufapi.dto;
-
-public interface IResponse {}
