+++ backend/src/main/java/net/archasmiel/skufapi/exception/AuthException.java
@@ -2,7 +2,7 @@ package net.archasmiel.skufapi.exception;
 
 public class AuthException extends Exception {
 
-  public AuthException(String details) {
-    super("Verification exception: " + details);
-  }
+    public AuthException(String details) {
+        super("Verification exception: " + details);
+    }
 }
