--- backend/src/main/java/net/archasmiel/skufapi/exception/GoogleTokenVerificationException.java
@@ -1,12 +1,0 @@
-package net.archasmiel.skufapi.exception;
-
-public class GoogleTokenVerificationException extends Exception {
-
-  public GoogleTokenVerificationException(String details) {
-    super("Critical exception in Google token verification: " + details);
-  }
-
-  public GoogleTokenVerificationException(GoogleTokenVerificationException e) {
-    super(e.getMessage());
-  }
-}
