+++ backend/src/main/java/net/archasmiel/skufapi/exception/RSAKeyException.java
@@ -2,7 +2,7 @@ package net.archasmiel.skufapi.exception;
 
 public class RSAKeyException extends RuntimeException {
 
-  public RSAKeyException(String details) {
-    super("RSA Key Exception: " + details);
-  }
+    public RSAKeyException(String details) {
+        super("RSA Key Exception: " + details);
+    }
 }
