--- backend/src/main/java/net/archasmiel/skufapi/model/Role.java
@@ -1,6 +1,0 @@
-package net.archasmiel.skufapi.model;
-
-public enum Role {
-  ROLE_USER,
-  ROLE_ADMIN,
-}
