+++ backend/src/main/java/net/archasmiel/skufapi/security/SecurityArgumentResolver.java
@@ -1,0 +1,32 @@
+package net.archasmiel.skufapi.security;
+
+import net.archasmiel.skufapi.api.response.ApiResponse;
+import net.archasmiel.skufapi.api.response.auth.MeResponse;
+import net.archasmiel.skufapi.exception.AuthException;
+import org.springframework.core.MethodParameter;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.stereotype.Component;
+import org.springframework.web.bind.support.WebDataBinderFactory;
+import org.springframework.web.context.request.NativeWebRequest;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.method.support.ModelAndViewContainer;
+
+@Component
+public class SecurityArgumentResolver implements HandlerMethodArgumentResolver {
+
+    @Override
+    public boolean supportsParameter(MethodParameter parameter) {
+        return parameter.hasParameterAnnotation(SecurityContext.class)
+                && Authentication.class.isAssignableFrom(parameter.getParameterType());
+    }
+
+    @Override
+    public Object resolveArgument(
+            MethodParameter parameter, ModelAndViewContainer mavContainer,
+            NativeWebRequest webRequest, WebDataBinderFactory binderFactory) {
+        return SecurityContextHolder.getContext().getAuthentication();
+    }
+}
