+++ backend/src/main/java/net/archasmiel/skufapi/security/SecurityContext.java
@@ -5,4 +5,5 @@ import java.lang.annotation.*;
 @Target(ElementType.PARAMETER)
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
-public @interface SkufSecurity {}
+public @interface SecurityContext {
+}
