--- backend/src/main/java/net/archasmiel/skufapi/security/SkufSecurityArgumentResolver.java
@@ -1,30 +1,0 @@
-package net.archasmiel.skufapi.security;
-
-import org.springframework.core.MethodParameter;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.stereotype.Component;
-import org.springframework.web.bind.support.WebDataBinderFactory;
-import org.springframework.web.context.request.NativeWebRequest;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.method.support.ModelAndViewContainer;
-
-@Component
-public class SkufSecurityArgumentResolver implements HandlerMethodArgumentResolver {
-
-  @Override
-  public boolean supportsParameter(MethodParameter parameter) {
-    return parameter.hasParameterAnnotation(SkufSecurity.class)
-        && Authentication.class.isAssignableFrom(parameter.getParameterType());
-  }
-
-  @Override
-  public Object resolveArgument(
-      MethodParameter parameter,
-      ModelAndViewContainer mavContainer,
-      NativeWebRequest webRequest,
-      WebDataBinderFactory binderFactory)
-      throws Exception {
-    return SecurityContextHolder.getContext().getAuthentication();
-  }
-}
