+++ backend/src/main/resources/application.properties
@@ -21,5 +21,4 @@ spring.security.oauth2.client.provider.google.issuer-uri=https://accounts.google
 
 # JWT
 # Keys are in folder "keys/"
-key.secret.frontend=${SECRET_FRONTEND}
 jwt.token.expiration=${JWT_EXPIRATION_MS}
